Split each interaction into its own JAX-RS resource
